Refactor parse_event
