+++ api/get_game.py
@@ -6,26 +6,7 @@ import json
 import decimal
 from shared.response import * 
 from shared.db import * 
-
-
-def parse_event(event):
-    # Basic input validation
-    if not isinstance(event, dict):
-        return False
-
-    # Handle both direct triggers and API Gateway
-    body = event.get("body", event)
-    if isinstance(body, str):
-        try:
-            body = json.loads(body)
-        except ValueError:
-            return None
-    path_params = event.get("pathParameters", {})
-    query_params = event.get("queryStringParameters", {})
-    body.update(path_params)
-    body.update(query_params)
-    return body
-
+from shared.api_gateway import parse_event
 
 def is_valid_uuid(value):
     try:
