+++ api/invite-aiagent.py
@@ -7,28 +7,10 @@ from boto3.dynamodb.conditions import Key
 import decimal
 from shared.response import * 
 from shared.db import *
+from shared.api_gateway import parse_event
 
 
 AGENT_MAPPING = json.loads(os.environ.get("agent_mapping"))
-
-
-def parse_event(event):
-    # Basic input validation
-    if not isinstance(event, dict):
-        return False
-
-    # Handle both direct triggers and API Gateway
-    body = event.get("body", event)
-    if isinstance(body, str):
-        try:
-            body = json.loads(body)
-        except ValueError:
-            return None
-    path_params = event.get("pathParameters", {})
-    query_params = event.get("queryStringParameters", {})
-    body.update(path_params)
-    body.update(query_params)
-    return body
 
 
 def is_valid_uuid(value):
