+++ api/send-reaction.py
@@ -12,6 +12,7 @@ logger.setLevel(logging.INFO)
 import re
 from shared.response import * 
 from shared.db import *
+from shared.api_gateway import parse_event
 
 
 sqs_client = boto3.client("sqs")
@@ -48,25 +49,6 @@ def send_queue_message(message):
                                 MessageBody=json.dumps(message, cls=DecimalEncoder))
     except ClientError:
         return
-
-
-def parse_event(event):
-    # Basic input validation
-    if not isinstance(event, dict):
-        return False
-
-    # Handle both direct triggers and API Gateway
-    body = event.get("body", event)
-    if isinstance(body, str):
-        try:
-            body = json.loads(body)
-        except ValueError:
-            return None
-    path_params = event.get("pathParameters", {})
-    query_params = event.get("queryStringParameters", {})
-    body.update(path_params)
-    body.update(query_params)
-    return body
 
 
 def is_valid_uuid(value):
