+++ api/shared/api_gateway.py
@@ -1,0 +1,23 @@
+import json
+
+def parse_event(event):
+    # Basic input validation
+    if not isinstance(event, dict):
+        if isinstance(event, str):
+            return json.loads(event)
+        return False
+
+    # Handle both direct triggers and API Gateway
+    body = event.get("body", event)
+    if isinstance(body, str):
+        try:
+            body = json.loads(body)
+        except ValueError:
+            return None
+    path_params = event.get("pathParameters", {})
+    query_params = event.get("queryStringParameters", {})
+    body.update(path_params)
+    body.update(query_params)
+    return body
+
+
