+++ api/watch-game-stream.py
@@ -10,6 +10,7 @@ import logging
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 from shared.response import * 
+from shared.api_gateway import parse_event
 
 
 sqs_client = boto3.client("sqs")
@@ -62,27 +63,6 @@ def is_valid_uuid(value):
         return True
     except ValueError:
         return False
-
-
-def parse_event(event):
-    # Basic input validation
-    if not isinstance(event, dict):
-        if isinstance(event, str):
-            return json.loads(event)
-        return False
-
-    # Handle both direct triggers and API Gateway
-    body = event.get("body", event)
-    if isinstance(body, str):
-        try:
-            body = json.loads(body)
-        except ValueError:
-            return None
-    path_params = event.get("pathParameters", {})
-    query_params = event.get("queryStringParameters", {})
-    body.update(path_params)
-    body.update(query_params)
-    return body
 
 
 def find_connected_players(game):
